add custom tilte bar + code clean up

+++ engine/include/utility/screen.hpp
@@ -1,5 +1,5 @@
 #include <string>
 
 namespace Anterle {
-void saveScreenshotToFile(std::string filename, int windowWidth, int windowHeight);
+void saveScreenshotToFile(std::string filename, uint16_t windowWidth, uint16_t windowHeight);
 }
